AI-survey-creator-tool-main/Survey/requirements.txt
flask
flask-login
flask-bcrypt
python-dotenv
requests
transformers
torch
spacy
webauthn
pydantic
deep_translator
twilio
geocoder
qrcode[pil]
dateutils